requirements/constraints.txt
# Everything shared (NO torch/torchvision/torchaudio here)
accelerate==0.31.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
controlnet-aux==0.0.6
cycler==0.12.1
diffusers==0.29.2
einops==0.7.0
filelock==3.13.1
fonttools==4.59.1
fsspec==2024.6.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
kiwisolver==1.4.9
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.1
peft==0.11.1
pillow==10.4.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.15.3
six==1.17.0
sympy==1.13.1
tifffile==2025.5.10
timm==0.9.16
tokenizers==0.19.1
tqdm==4.67.1
transformers==4.43.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0

# OpenCV portability (headless on non-Windows; full contrib on Windows)
opencv-contrib-python-headless==4.12.0.88; platform_system != "Windows"
opencv-contrib-python==4.12.0.88; platform_system == "Windows"
